// trustee-frontend/src/app/model/ImmunizationReport.ts
export interface ImmunizationReport {
    interesovanjeNum: number;
    zahtevNum: number;
    sertifikatNum: number;
    ukupnoDatoNum: number;
    doza1Num: number;
    doza2Num: number;
    doza3Num: number;
    pfizerNum: number;
    sinopharmNum: number;
    sputnikNum: number;
    astraNum: number;
}